Revert "Let cargo test work (#96)"
This reverts commit 34d445cec317409ae099b7602efecd6aa7ff7b28.
