+++ savvy-bindgen/src/gen/templates/Cargo_toml
@@ -8,6 +8,3 @@ crate-type = ["staticlib"]
 
 [dependencies]
 savvy = "*"
-
-[dev-dependencies]
-savvy = {{ version = "*", features = ["fake-libR"] }}
