+++ savvy-ffi/Cargo.toml
@@ -8,8 +8,5 @@ license.workspace = true
 homepage.workspace = true
 repository.workspace = true
 
-[features]
-fake-libR = []
-
 [package.metadata.dist]
 dist = false
